WIP: Generate client-side routes during build
Pulled in latest react-router and setup a basic `/home` and `/about` routes. Each route has its own hook to get initialization data for that specific route(s). This initialization data is dumped on each route/page combo to seed store state.

Using babel-node to run webpack in order to retrieve client-side routes at build time. Routes contain references to its respective component, so it was not sufficient to simply run webpack in node without first processing it with babel; hence the need for babel-cli and babel-node.